+++ Code/Tools/FBuild/FBuildCore/WorkerPool/Job.cpp
@@ -3,14 +3,13 @@
 
 // Includes
 //------------------------------------------------------------------------------
-#include "Tools/FBuild/FBuildCore/PrecompiledHeader.h"
-
 #include "Job.h"
 
 #include "Tools/FBuild/FBuildCore/Graph/Node.h"
 #include "Tools/FBuild/FBuildCore/FLog.h"
 
 #include "Core/Env/Assert.h"
+#include "Core/FileIO/FileIO.h"
 #include "Core/FileIO/IOStream.h"
 #include "Core/Process/Atomic.h"
 #include "Core/Profile/Profile.h"
@@ -78,7 +77,7 @@ void Job::OwnData( void * data, size_t size, bool compressed )
         if ( m_IsLocal )
         {
             ASSERT( s_TotalLocalDataMemoryUsage >= m_DataSize );
-            AtomicSub64( &s_TotalLocalDataMemoryUsage, m_DataSize );
+            AtomicSub64( &s_TotalLocalDataMemoryUsage, (int32_t)m_DataSize );
         }
     }
 
@@ -90,7 +89,7 @@ void Job::OwnData( void * data, size_t size, bool compressed )
     // Update total memory use tracking
     if ( m_IsLocal )
     {
-        AtomicAdd64( &s_TotalLocalDataMemoryUsage, m_DataSize );
+        AtomicAdd64( &s_TotalLocalDataMemoryUsage, (int32_t)m_DataSize );
     }
 }
